src/pyproject.toml
[tool.poetry]
name = "terminal-hero"
version = "1.0.0"
description = "AI-Powered Terminal Troubleshooting Agent with Multi-Agent System"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
langgraph = ">=0.0.20"
langchain-openai = ">=0.0.5"
openai = ">=1.10.0"
typer = ">=0.9.0"
rich = ">=13.7.0"
textual = ">=0.48.0"
pydantic = ">=2.5.0"
psutil = ">=5.9.0"
distro = ">=1.9.0"
duckduckgo-search = ">=4.1.0"
requests = ">=2.31.0"
beautifulsoup4 = ">=4.12.0"
lxml = ">=5.1.0"
python-dotenv = ">=1.0.0"

[tool.poetry.dev-dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
ruff = "^0.1.0"
mypy = "^1.7.0"

[tool.poetry.scripts]
terminal-hero = "src.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false